requirements.txt
Django==1.8.4
django-bootstrap-form==3.2
django-debug-toolbar==1.3.2
django-disqus==0.5
django-forms-bootstrap==3.0.1
Pillow==3.0.0
python-dateutil==2.4.2
requests==2.8.0
sqlparse==0.1.16
stripe==1.24.1
virtualenv==13.1.2
wheel==0.24.0
mysql-python
django-tinymce==2.0.4